Update routers to make it more flexible
